+++ CHANGELOG.md
@@ -14,6 +14,9 @@
   - 3.0+ version of the ESP32 Arduino Core now required.
   - USB-CDC (aka native USB) appears to be broken in the 3.0 core. Will eventually hang if sending a lot of data both directions at the same time. Use one of the standard UART interfaces until this is fixed.
 
+- Raspberry Pi Pico (RP2040)
+  - WS2812 LED strips work now.
+
 ### New Peripherals
 
 - ADS1100 Analog-to-Digital Converter: